add a script to generate .py files from .ipynb docs/source/examples
