+++ examples/partitioner_plots_example.py
@@ -1,30 +1,33 @@
-"""
+#!/usr/bin/env python
+# coding: utf-8
 
-Partitioner Plot Example
--------------------------
+# # Partitioner Plot Example
 
-This example shows how to visualize the tree-based partitioning performed by 
-KDQTreePartitioner.
+# These examples show how to visualize the tree-based partitioning performed by 
+# KDQTreePartitioner.
+# 
+# KDQTreePartitioner takes input data and partitions the data into a kdq-Tree. New 
+# data can be passed subsequently and counted according to that same reference 
+# partition. The two datasets can then be compared.
+# 
+# This partitioner is used within the library within ``data_drift.kdq_tree``, 
+# where the reference data is used to build the partition, and the test data is 
+# then divided by that partition, in order to compare the two distributions.
 
-KDQTreePartitioner takes input data and partitions the data into a kdq-Tree. New 
-data can be passed subsequently and counted according to that same reference 
-partition. The two datasets can then be compared.
+# In[ ]:
 
-This partitioner is used within the library within ``data_drift.kdq_tree``, 
-where the reference data is used to build the partition, and the test data is 
-then divided by that partition, in order to compare the two distributions.
 
-"""
+## Imports ##
 
 import numpy as np
 import plotly.express as px
-
 from menelaus.partitioners.KDQTreePartitioner import KDQTreePartitioner, KDQTreeNode
 
 
-################################################################################
-################################ Basic Plot ####################################
-################################################################################
+# ## Basic Plot Example
+
+# In[ ]:
+
 
 # Create some data and build the tree.
 data = np.random.randint(0, 10, (20, 3))
@@ -35,12 +38,9 @@ KDQTreeNode.as_text(root)
 # Grab the output for plotly.
 df_plot = kp.to_plotly_dataframe(tree_id1="build")
 
-# This uses only cell_count, which depends only one tree, and doesn't
-# compare them.
 
-# Note that it doesn't include any special statistics, just the count.
-# df_plot
-df_plot.to_csv("example_partitioner_plots_basic.csv", index=False)
+# In[ ]:
+
 
 # Note that plotly's textinfo value could be used to add alternative values,
 # if desired.
@@ -52,8 +52,12 @@ fig = px.treemap(
     values="cell_count",
 )
 fig.update_traces(root_color="lightgrey")
-# fig.show()
-fig.write_html(f"example_partitioner_plots_basic_plot.html")
+fig.show()
+# fig.write_html(f"example_partitioner_lots_basic_plot.html")
+
+
+# In[ ]:
+
 
 ############## Filter by depth ##############
 kp = KDQTreePartitioner(count_ubound=25)
@@ -72,13 +76,14 @@ fig = px.treemap(
     values="cell_count",
 )
 fig.update_traces(root_color="lightgrey")
-# fig.show()
-fig.write_html(f"example_partitioner_plots_basic_plot_depth.html")
+fig.show()
+# fig.write_html(f"example_partitioner_lots_basic_plot_depth.html")
 
 
-################################################################################
-################################# Modifications ################################
-################################################################################
+# ## Modifications Example
+
+# In[ ]:
+
 
 # Count differences between builds can be accessed.
 # This can be used to modify the display.
@@ -88,8 +93,10 @@ df2 = np.random.sample([50, 3])
 _ = kp.build(df)
 _ = kp.fill(df2, "fill1")
 df_plot = kp.to_plotly_dataframe("build", "fill1")
-# df_plot
-df_plot.to_csv("example_partitioner_plots_modifications.csv", index=False)
+
+
+# In[ ]:
+
 
 fig = px.treemap(
     data_frame=df_plot,
@@ -99,8 +106,11 @@ fig = px.treemap(
     color="count_diff",
 )
 fig.update_traces(root_color="lightgrey")
-# fig.show()
-fig.write_html(f"example_partitioner_plots_modifications1_count.html")
+fig.show()
+# fig.write_html(f"example_partitioner_lots_modifications1_count.html")
+
+
+# In[ ]:
 
 
 ############## Display additional information ##############
@@ -114,8 +124,11 @@ fig = px.treemap(
 fig.update_traces(
     root_color="lightgrey", textinfo="label+current path"
 )  # see textinfo in https://plotly.com/python/reference/treemap/
-# fig.show()
-fig.write_html(f"example_partitioner_plots_modifications2_path.html")
+fig.show()
+# fig.write_html(f"example_partitioner_lots_modifications2_path.html")
+
+
+# In[ ]:
 
 
 ##### Access the plot and color using the Kulldorff Spatial Scan Statistic (KSS) #####
@@ -130,8 +143,11 @@ fig = px.treemap(
     color_continuous_scale="blues",
 )
 fig.update_traces(root_color="lightgrey")
-# fig.show()
-fig.write_html(f"example_partitioner_plots_modifications3_kss.html")
+fig.show()
+# fig.write_html(f"example_partitioner_lots_modifications3_kss.html")
+
+
+# In[ ]:
 
 
 ############# Outline the cells according to the direction of change in counts #############
@@ -155,13 +171,15 @@ fig.update_traces(
     },
     root_color="lightgrey",
 )
-# fig.show()
-fig.write_html(f"example_partitioner_plots_modifications4_outline.html")
+fig.show()
+# fig.write_html(f"example_partitioner_lots_modifications4_outline.html")
 
 
-################################################################################
-################################# Alternatives #################################
-################################################################################
+# ## Alternatives Example
+
+# In[ ]:
+
+
 # For the most part, sunburst and icicle plots take the same arguments and
 # behave, though they're not as space-efficient visually.
 
@@ -177,6 +195,10 @@ df_plot.loc[df_plot.count_diff < 0, "count_dir"] = "red"
 df_plot.loc[df_plot.count_diff == 0, "count_dir"] = "lightgrey"
 df_plot.loc[df_plot.count_diff > 0, "count_dir"] = "green"
 
+
+# In[ ]:
+
+
 ############# Sunburst Plot #############
 fig = px.sunburst(
     data_frame=df_plot,
@@ -190,8 +212,11 @@ fig.update_traces(
     insidetextfont={"color": df_plot.count_dir},
     root_color="lightgrey",
 )
-# fig.show()
-fig.write_html(f"example_partitioner_plots_alternatives_sunburst.html")
+fig.show()
+# fig.write_html(f"example_partitioner_lots_alternatives_sunburst.html")
+
+
+# In[ ]:
 
 
 ############# Icicle Plot #############
@@ -207,5 +232,6 @@ fig.update_traces(
     insidetextfont={"color": df_plot.count_dir},
     root_color="lightgrey",
 )
-# fig.show()
-fig.write_html(f"example_partitioner_plots_alternatives_icicle.html")
+fig.show()
+# fig.write_html(f"example_partitioner_lots_alternatives_icicle.html")
+
